fix: rollup typescript error
